# floppa-web-shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Floppa VPN Admin API",
    "description": "API for Floppa VPN admin panel and user management",
    "version": "0.1.0"
  },
  "paths": {
    "/admin/avatars": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Batch-fetch cached avatars as data URLs (admin). Returns only users that have a cached avatar,\nkeyed by user id (string). Avoids one request per row in the admin user list.",
        "operationId": "get_avatars_batch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AvatarBatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Map of user id → data URL",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "propertyNames": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Not an admin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/admin/users/{id}/avatar": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get a user's cached avatar (admin).",
        "operationId": "get_user_avatar",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Avatar image bytes",
            "content": {
              "image/jpeg": {}
            }
          },
          "304": {
            "description": "Not modified"
          },
          "403": {
            "description": "Not an admin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "No avatar cached",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/me/avatar": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Get the current user's cached avatar.",
        "operationId": "get_my_avatar",
        "responses": {
          "200": {
            "description": "Avatar image bytes",
            "content": {
              "image/jpeg": {}
            }
          },
          "304": {
            "description": "Not modified"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "No avatar cached",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "AvatarBatchRequest": {
        "type": "object",
        "required": [
          "user_ids"
        ],
        "properties": {
          "user_ids": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          }
        }
      },
...
    }
  }
}